refactor(memory): split the core service and MCP helpers along their seams
Pure moves, no behavior change. src/core/memory.ts (960 LOC) becomes a
facade over memory-types (contracts, constants, MemoryError), memory-fence
(validation, context fence, receipts, record materialization),
memory-remember (candidates, add, supersede under the lease), and
memory-recall (retrieval legs, RRF fusion, budget). Every existing
`src/core/memory` import path keeps working through re-exports.

The MCP helpers shared by gno_recall and gno_remember (identity mapping,
service construction, error rethrow, scopes schema) move from
memory-recall.ts to memory-shared.ts. The buildFts5Query grammar JSDoc in
the SQLite adapter now sits over buildFts5Query again instead of
SUPERSEDED_EXCLUSION_SQL.
